LayoutTests/fast/dom/shadow/shadowdom-for-form-associated-element-useragent-expected.txt

When a AuthorShadowRoot is added, UserAgentShadowRoot should exists as the oldest Shadow Root.

PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS oldestShadowRoot is not shadowRoot
PASS youngerShadowRoot is shadowRoot
PASS successfullyParsed is true

TEST COMPLETE
